Finishing: authorization
Project done

+++ app/Http/Controllers/ListingController.php
@@ -30,6 +30,7 @@ public function create() {
     public function store(Request $request) {
         //dd($request->all());
         //dd($request->file('logo'));
+        //dd(auth()->id());
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
@@ -40,10 +41,16 @@ public function create() {
             'description' => 'required',
         ]); 
 
+        // if (auth()->check()) {
+        //     $formFields['user_id'] = auth()->id();
+        // }
+
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+        
         Listing::create($formFields);
         //For a message, create a component like flash-message.blade.php
         return redirect('/')->with('message', 'Listing created successfuly');
@@ -53,11 +60,22 @@ public function create() {
     public function edit(Listing $listing) {
         //dd($listing);
         //dd($listing->title);
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
     //Update Listing data
     public function update(Request $request, Listing $listing) {
+
+            //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -78,7 +96,17 @@ public function create() {
 
     //Destroy Method
     public function destroy(Listing $listing) {
+
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    //Manage Listing
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
